feat: upgrade Solana vault swaps (#5845)
* chore: PoC

* docs: nit

* chore: add upgrade program to upgrade test

* chore: update step id

* chore: update tests

* chore: add solana --version print

* chore: add solana path

* chore: improve bouncer and upgrade

* chore: run upgrade test

* chore: update to v1.1.0-rc0 except image

* chore: add test upgrade_program transaction builder

* chore: restore comment format

* feat: update rpc to take in seed instead of swap_event_account

* chore: lint bouncer

* chore: revert upgrade flag

* feat: add support for spill_address

* chore: address PR comments

* chore: update to sol-contracts tag 1.1.0

* chore: fix full bouncer ci

+++ foreign-chains/solana/sol-prim/src/consts.rs
@@ -23,7 +23,7 @@ pub const SOLANA_DIGEST_LEN: usize = 32;
 
 // NB: this includes the bump-seed!!!
 pub const SOLANA_PDA_MAX_SEEDS: u8 = 16;
-pub const SOLANA_PDA_MAX_SEED_LEN: usize = 32;
+pub const SOLANA_PDA_MAX_SEED_LEN: u32 = 32;
 pub const SOLANA_PDA_MARKER: &[u8; 21] = b"ProgramDerivedAddress";
 
 pub const HASH_BYTES: usize = 32;
